+++ CHANGELOG.md
@@ -1,9 +1,11 @@
+## 1.3.0
+* add: <code>vue2-ts</code> options
 ## 1.2.0
 * add: <code>vue3</code> options
 * add: command alias: <code>npx veph</code>
 * add: jsconfig.json in vue2
-* add: vscode extensiton(<code>octref.vetur</code>) when select vue2
-* add: vscode extensiton(<code>johnsoncodehk.volar</code>) when select vue3
+* add: vscode extension(<code>octref.vetur</code>) when select vue2
+* add: vscode extension(<code>johnsoncodehk.volar</code>) when select vue3
 * change: <code>vue</code> to <code>vue2</code>
 ## 1.1.2
 * add: add <code>js</code> option, as default selection
@@ -11,7 +13,7 @@
 * change: sync code
 * change: single select env
 * change: every env has it's eslintrc.xx.js
-* change: remove <code>vue</code> as default chioce
+* change: remove <code>vue</code> as default choice
 * change: remove useless rules in <code>.eslintrc.js</code>
 ## 1.1.1
 * fix: update vscode setting.json not work